#109 have dropdowns have a pointer

+++ db/schema.rb
@@ -51,8 +51,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_04_14_145051) do
   end
 
   create_table "project_tags", force: :cascade do |t|
-    t.integer "project_log_id", null: false
-    t.integer "tag_id", null: false
+    t.bigint "project_log_id", null: false
+    t.bigint "tag_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["project_log_id", "tag_id"], name: "index_project_tags_on_project_log_id_and_tag_id", unique: true
